move cb to custom
