+++ internal/custom/custom_test.go
@@ -1,36 +1,22 @@
-package auth0
+package custom
 
 import (
 	"bytes"
 	"encoding/json"
+	"gopkg.in/auth0.v4"
 	"testing"
 )
 
-func TestConvertibleBool(t *testing.T) {
-	for _, test := range []struct {
-		in       *ConvertibleBoolean
-		expected ConvertibleBoolean
-	}{
-		{nil, false},
-		{ConvertibleBool(false), false},
-		{ConvertibleBool(true), true},
-	} {
-		have := ConvertibleBoolValue(test.in)
-		if have != test.expected {
-			t.Errorf("unexpected output. have %v, expected %v", have, test.expected)
-		}
-	}
-}
 
 func TestConvertibleBoolean_UnmarshalJSON(t *testing.T) {
 	for _, test := range []struct {
 		in       string
 		expected *ConvertibleBoolean
 	}{
-		{`{"bool": false}`, ConvertibleBool(false)},
-		{`{"bool": "false"}`, ConvertibleBool(false)},
-		{`{"bool": true}`, ConvertibleBool(true)},
-		{`{"bool": "true"}`, ConvertibleBool(true)},
+		{`{"bool": false}`, auth0.ConvertibleBool(false)},
+		{`{"bool": "false"}`, auth0.ConvertibleBool(false)},
+		{`{"bool": true}`, auth0.ConvertibleBool(true)},
+		{`{"bool": "true"}`, auth0.ConvertibleBool(true)},
 	} {
 		var ts struct {
 			Bool *ConvertibleBoolean `json:"bool,omitempty"`
@@ -40,7 +26,7 @@ func TestConvertibleBoolean_UnmarshalJSON(t *testing.T) {
 			t.Errorf("expected no error, got %s", err)
 		}
 
-		if ConvertibleBoolValue(ts.Bool) != ConvertibleBoolValue(test.expected) {
+		if auth0.ConvertibleBoolValue(ts.Bool) != auth0.ConvertibleBoolValue(test.expected) {
 			t.Errorf("unexpected output. have %v, expected %v", ts.Bool, test.expected)
 		}
 	}
@@ -51,8 +37,8 @@ func TestConvertibleBoolean_MarshalJSON(t *testing.T) {
 		in       *ConvertibleBoolean
 		expected []byte
 	}{
-		{ConvertibleBool(false), []byte(`{"bool":false}`)},
-		{ConvertibleBool(true), []byte(`{"bool":true}`)},
+		{auth0.ConvertibleBool(false), []byte(`{"bool":false}`)},
+		{auth0.ConvertibleBool(true), []byte(`{"bool":true}`)},
 	} {
 		var ts struct {
 			Bool *ConvertibleBoolean `json:"bool,omitempty"`
