+++ management/user.go
@@ -1,9 +1,8 @@
 package management
 
 import (
-	"gopkg.in/auth0.v3"
-
 	"encoding/json"
+	"gopkg.in/auth0.v4/internal/custom"
 	"reflect"
 	"strconv"
 	"time"
@@ -64,7 +63,7 @@ type User struct {
 	// True if the user's email is verified, false otherwise. If it is true then
 	// the user will not receive a verification email, unless verify_email: true
 	// was specified.
-	EmailVerified *auth0.ConvertibleBoolean `json:"email_verified,omitempty"`
+	EmailVerified *custom.ConvertibleBoolean `json:"email_verified,omitempty"`
 
 	// If true, the user will receive a verification email after creation, even
 	// if created with email_verified set to true. If false, the user will not
